Changed build system to Gnu make and reorganized the directory structure.
I discovered that Rant has been orphaned; the main developers have all
quit to work on other projects since Rake seems to be winning in the
Ruby space.  Since it is still early in the project, I decided to
switch to gmake.  I had been considering this anyway because of the
nice feature of parallel builds.

I took advantage of the change to move all source files in to src and
all build files into obj.  The executables are in the root.