Separate declaration of mockable stringstream and Makefile targets for mock and normal executables
I separated the mockable stringstream extraction operator into a
non-member function (since otherwise, the superclass member overloads
were not recognized).  To do this, I had to move the code to a .cpp
file because of linkage.  This meant that I'd have to rebuild the
library with the appropriate version of stringstream.  Thus, I had to
introduce new makefile targets for mock-linked executables and
normally-linked executables.