Simplified the use of Key objects by creating a KeySptr class with its own "<"
This allowed me to get rid of the DereferenceLess class and make the
KeyPair class into a mere typedef.  It also simplified the definitions
of the sets of keys.

I didn't just define an operator < on the Key* objects because the
natural semantics of those is to directly compare the pointers.  The
KeySptr class allows me to distinguish between pointers and pointers
intended as standins for actual objects that are subclasses of Key.