Buggy attempt to fix bug that identifications are not removed when their underlying peak is removed.
Now, clicking on an identified peak can result in just the identification being removed.

There may be a bug that removes peaks in other spectra/bins.

I think that if all the peaks in a bin are removed then the program
thinks that it needs to recalculate the peak locations.  In fact, I'm
pretty sure now.