Replaced all instances of get_cur_peaks and set_peaks with new routines and adapted the surrounding code to match the new semantics of peaks being all the peaks in the spectrum.
Fixed two bugs in bins_affected_by_peak_change - it wasn't detecting
changes at the edge of the bin and it returned indices into a matrix
rather than a vector.

Fixed bug in set_spectrum_peaks_no_gui - it didn't work when no bins were affected by the change.

Changed update peaks button to be longer and to display the message "Recalculate Peaks" when the deconvolution is up-to-date and "Update Peaks" when it is not up-to-date.

Changed to always mark the current deconvolution as up-to-date even if
it changes the peak locations.  This means that unless the user
changes the peaks, the last deconvolution the user saw will be the one
used for the output.