Added cleaned version of BinMap_Sept_2011_Reo.xlsx.  I did this in ods because my copy of openoffice can't save xlsx files.
I added a field for whether the metabolite is known (this will allow
for the possibility that the id will be changed in the future, a
metabolite may have at most one known metabolite ID but many ids for
unknown compounds may point to it as well.)

I renamed the # field as ID fields since that gives more a sense of a
key and it is not a peak # but a bin id (since a we may have a doublet
or a triplet, which contain more than one peak).

I added an Expected Peaks field because trigonelline usually has 3
peaks because the shifts are such that the two doublets overlap to
create what looks like a triplet.  This will keep the user from having
to continually affirm that they really only want to identify 3 peaks
in that area.

I added a notes field for explaining details like the above to future
users of the metab-map

I fixed some bins which contained more than one compound.  They should
be two separate identical bins.  Otherwise, the user does not know
what to identify

I also fixed some of the formatting in the metabolite names.