Hand merged original shared metab-map in old format with reo's new map from earlier this month.  This will keep bin numbers universal since we are going to continue using the bin number as a traditional primary key.
The merge involved a lot of adding information and checking as well.
I added hmdb numbers for many compounds as well as j values and notes
on possible inaccuracies.

The bins that were in the new map but not the old one do not have bin
numbers yet.  I'll do that on my next checkin.

This took ALL DAY. AAARG!

Issue #68