Added failing test cases to test whether eq works correctly with sample_types field.
Decided that, since the sample_types field is really a set, the
sample_types field should be sorted so that equality was the same as
set-equality.  Changed some test-cases to conform to this constraint.

Added some new spaces to the
testCompoundBinConstructorSampleTypesFiveTypes routine to ensure that
spaces were really getting removed on the input parsing.

Issue #100