I ignored the cbfs_lib file.
This is necessary because I am currently using MINE.jar under the
hood.  The license does not appear to give me the right to repackage
it in my own jar.  So, I need to have the libraries external and in
their own directory when exporting.  When I remove the dependence on
MINE.jar, then this should go away.