Changed the interface for Relation objects. Instead of mainly being concerned with returning a single sample, they are now expected to return a whole instance worth of samples.
This simplifies code in several ways and makes it more
efficient. First, code is more efficient because much less object
creation (no little sample objects). Second, the instance class was
designed to be almost directly plug-in compatible with the mic code,
so no code complexity to convert from one format to another and no
extra program time spent on the conversion. Third, there is no
dependence between the samples method and sample method of the
Relation class, so no problem keeping their return values consistent
or specifying in the documentation that they must be consistent.

The downside is that every relation class will have code to create
arrays and loop through, filling them.