Changed parameters to match my thoughts on the control urine.
Don't display output, vary the pH. Use maximum double as the SNR (so
as close to 0 as I can get without changing the code). Set the number
of replicates to 200 (and none for mix 2), ranged the x-axis from -1
to 12 and 32k data points.

pH is now 7.6 +/- 0.15. Reo said they buffer (so not natural ph of
5-6) and that probably closer to 8 than 7.

Changed the names of the input and output files.