Fixed several bugs. (As committed, things didn't work.)
true_dilutions should have been a row vector not a column vector.

use_bin should have been a single logical vector not a cell array with
a logical vector in it.

The number of spectra is the number of columns in the Y, not the size
of the spectral collections cell array.

Added the other implemented methods to the list of methods to run.

Also, the estimated time to run is up to 87 minutes.