Added a file which annotates the normal urine file with:
* The directory information/experiment number for the metabolite
spectrum in NSSD

* what I believe to be the HMDB name

* HMDB ID (also known as the Accession Number)

This was mostly done automatically using the unix sort and join
commands (after some emacs search and replace to make everything TSV
format) but due to an inconsistency in the naming, the last three
entries had to be matched by hand. However, the matches were obvious,
so I think there was no error inserted.

Doing this will allow me to include this information in the spectra
that I generate from the NSSD files.

Note: the local names file has an entry for glyceric acid, but no such
entry appears in the NSSD.