The positive residual constraint is not working well enough to deconvolve large bins. Smaller peaks are getting swamped by the sum of many tails of large peaks because of fitting each point only on its local neighborhood.
I've gotten rid of the positive residual constraint (and changed the
height to be the last height).

Here is one attempt to replace the positive residual constraint. (This
attempt doesn't work, and I'm going to get rid of it, but I want a
record that I worked on it). In this attempt, I try rescaling all the
other peaks when I come to a marked peak that has a residual below
some multiple of the noise threshold (you can see in the source that I
tried several multiples and all were ineffective.)