Fixed bugs that were the result of not having been able to run the code before the last commit. This version takes FOREVER to run on a bin with lots of peaks - and doesn't produce very good results. My next idea is to do the same thing but start with the smallest identified peaks first - their tails won't drown out the bigger peaks. And, if they are identified, they should still have some local hump that will show their position and allow them to be fitted.
Bugs:

Forgot that glp was an ARRAY of peak objects, so had to adapt all
things to vector operations.

Needed to take care of the case where a only a single peak had been
identified. (I still need to also take care of the case where no peaks
are identified).

Added some test code to the inner fitting loop for unburying small
peaks.