Allow width to vary by up to 2x estimated in the case of single peaks rather than being forced to exactly what was estimated by the rough deconvolution.
Made height lower bound the estimated height when there is a single
peak in the bin. For bins with single, short peaks, the local
deconvolution will be more accurate because the global deconvolution
will try to fit the entire bin, taking up any bias in the noise by
lowering and widening the peak. Bounding the height in this way keeps
that from happening.