I was able to look at the missing figures. Now, I'll try to get them to display in general. I think the problem is that I was opening figures so fast that the window manager crashed opening some of them. But matlab still kept the handles even after I slowed down and was writing to the invalid handles. I'm going to try it slower with a fresh run of matlab that will be acquiring new window handles.
After trying it, it seems the window manager crashes on opeining
window 12 and has trouble again on later window openings. This time,
it seems to have stayed dead. (There are no window decorations
anymore.) Time for a reboot. It may be that the problem comes when I
do my maximize routine. I'll try it once more without maximizing the
windows.