Fixed bug in which the sampled distributions were not being equal-probability binned but equal-width binned. Added equal width binning for sampled distributions. Used the calculated counts to set the probabilities for the original sampled 7-bin histogram distribution (this makes it obvious that they are categorical distributions amenable to use with a dirichlet prior).
Added check that the updated version of the sampling code is being
used (checking for the existence of a variable that was only added in
most recent version).

Changed number of spectra used for sampled distributions to 5000 (I
found 7 bins that differed among the passes by more than 1% of
counts.)

Changed the number of steps per congestion to an explicit variable so
adding new steps in the future wouldn't be a problem (though it is
quite possible that these are the last steps I'm going to add)

Removed old equal-width calculations since I worked their content into
the main calculation.