Added code to calculate correlations in the original data with the data broken up by congestion.
Added code to calculate the same correlations in the deconvolved data
and code to print the results so I could look for problems with my
deconvolutions.

Added code to plot the joint distribution and joint rank distribution
of width and lorentzianness in an attempt to ferret out the
correlations detected. No patterns were revealed.