Removed test data for test_ExpDatum from the repository - it should be regenerated on each machine because the optimization is different for different machines. Added note to this effect.
Modified the censoring process when testing that update replaces
missing deconvolutions to remove particular types of deconvolutions
rather than particular indices. This makes the test cases a bit more
robust against future changes.

Changed assertDatumObjectsEqual to treat the deconvolutions field
specially so I can get better error messages.

Added an assertDatumObjectsApproxEqual method that may come in handy
in the future. (I wrote it when I thought that the differing
optimization results on different platforms were causing small enough
errors to make assertElementsAlmostEqual work, and kept it around
because it may be useful in the future)